requirements.txt
async-timeout==4.0.2
black==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.68.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-firestore==2.7.2
google-cloud-ndb==1.11.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.48.2
httplib2==0.21.0
idna==3.4
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
marshmallow-enum==1.5.1
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymemcache==4.0.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.28.1
rsa==4.9
six==1.16.0
tomli==2.0.1
twilio==7.16.0
typeguard==2.13.3
typing-inspect==0.8.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
